Data.Algorithm.Diff

Description

This is an implementation of the O(ND) diff algorithm as described in "An O(ND) Difference Algorithm and Its Variations (1986)" http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.4.6927. It is O(mn) in space. The algorithm is the same one used by standared Unix diff.

Documentation

type Diff a = PolyDiff a a

This is PolyDiff specialized so both sides are the same type.

data PolyDiff a b

A value is either from the First list, the Second or from Both. Both contains both the left and right values, in case you are using a form of equality that doesn't check all data (for example, if you are using a newtype to only perform equality on side of a tuple).

Constructors

First a

Second b

Both a b

Comparing lists for differences

getDiff :: Eq a => [a] -> [a] -> [Diff a]

Takes two lists and returns a list of differences between them. This is getDiffBy with == used as predicate.

getDiffBy :: (a -> b -> Bool) -> [a] -> [b] -> [PolyDiff a b]

A form of getDiff with no Eq constraint. Instead, an equality predicate is taken as the first argument.

Finding chunks of differences

getGroupedDiff :: Eq a => [a] -> [a] -> [Diff [a]]

Takes two lists and returns a list of differences between them, grouped into chunks. This is getGroupedDiffBy with == used as predicate.

getGroupedDiffBy :: (a -> b -> Bool) -> [a] -> [b] -> [PolyDiff [a] [b]]