Data.Algorithm.DiffOutput

Description

Generates a string output that is similar to diff normal mode

Documentation

diffToLineRanges :: [Diff [String]] -> [DiffOperation LineRange]

Converts Diffs to DiffOperations

ppDiff :: [Diff [String]] -> String

pretty print the differences. The output is similar to the output of the diff utility

prettyDiffs :: [DiffOperation LineRange] -> Doc

pretty print of diff operations

parsePrettyDiffs :: String -> [DiffOperation LineRange]

Parse pretty printed Diffs as DiffOperations

type LineNo = Int

Line number alias

data LineRange

Line Range: start, end and contents

Constructors

LineRange

Fields

lrNumbers :: (LineNo, LineNo)

lrContents :: [String]

data DiffOperation a

Diff Operation representing changes to apply

Constructors

Deletion a LineNo

Addition a LineNo

Change a a