Data.Algorithm.DiffContext

Description

Generates a grouped diff with merged runs, and outputs them in the manner of diff -u

Documentation

getContextDiff :: Eq a => Int -> [a] -> [a] -> ContextDiff a

Do a grouped diff and then split up the chunks into runs that contain differences surrounded by N lines of unchanged text. If there is less then 2N+1 lines of unchanged text between two changes, the runs are left merged.

prettyContextDiff

Arguments

:: Doc
Document 1 name

-> Doc
Document 2 name

-> (c -> Doc)
Element pretty printer

-> ContextDiff c

-> Doc

Pretty print a ContextDiff in the manner of diff -u.